feat: Integrate AdMob for rewarded ads and enhance token system
- Added google_mobile_ads dependency to support AdMob integration for rewarded ads.
- Implemented AdMobService to manage ad loading and displaying, allowing users to earn tokens by watching ads.
- Created RewardedAdWidget for displaying ads within the app, providing a seamless user experience.
- Updated token handling to support double values for token balance and transactions.
- Enhanced UI components to reflect changes in token management and ad interactions.
- Incremented app version to 1.1.0+15 in pubspec.yaml to reflect the latest updates.

+++ lib/core/services/in_app_purchase_service.dart
@@ -344,7 +344,7 @@ class InAppPurchaseService {
         // Grant tokens based on product ID
         final int? tokens = _productTokenMap[purchaseDetails.productID];
         if (tokens != null) {
-          await TokenService.addTokens(tokens);
+          await TokenService.addTokens(tokens as double);
           debugPrint(
             'Added $tokens tokens for purchase ${purchaseDetails.productID}',
           );
@@ -406,7 +406,7 @@ class InAppPurchaseService {
       try {
         final int? tokens = _productTokenMap[purchaseDetails.productID];
         if (tokens != null) {
-          await TokenService.addTokens(tokens);
+          await TokenService.addTokens(tokens as double);
           debugPrint(
             'Added $tokens tokens for restored purchase ${purchaseDetails.productID}',
           );
